requirements.txt
Pillow
scipy
geopy
requests
bs4
